Remove a number of integration tests which can be unit tests instead
Since each of these unit tests are quite large, removing just two of them would
result in a 1GB+ reduction in disk space utilization. This pull request removes
a few of the sets of tests which rely on tests/data/ but do not require any
non-default features.

Sponsored-by: Databricks Inc
